Fix build errors and warnings
This commit addresses several build issues:

1.  **BinderInterceptor Incomplete Type:**
    *   Moved the `BinderInterceptor` class definition from
        `binder_interceptor.cpp` to a new header file
        `binder_interceptor.h`.
    *   Included `binder_interceptor.h` in `binder_interceptor.cpp`.
    *   Made `gPropertyServiceBinder`, `handleIntercept`, and
        `needIntercept` members/methods public to allow access
        from other parts of the code.

2.  **Parcel API Errors:**
    *   Replaced `writeInterfaceToken` with `writeString16` for writing
        the interface token in `binder_interceptor.cpp`.
    *   Replaced `readCString` with `readString16` followed by a
        conversion to `String8` to get a `const char*`, ensuring
        `nullptr` is handled correctly for empty strings.
    *   Added `#include <utils/String8.h>` for `String8` usage.

3.  **Format Specifier Warnings (`elf_util.cpp`):**
    *   Corrected `snprintf` format specifiers to use `%llx` for
        `Elf64_Addr` and `Elf64_Xword` types, and `%llu` for
        `Elf64_Off` types.
    *   Added explicit C-style casts to `(unsigned long long)` for
        arguments in `snprintf` to ensure type safety with the
        new format specifiers.

4.  **Unused Variable Warning (`inject/utils.cpp`):**
    *   Added `[[maybe_unused]]` attribute to the unused loop
        variable `a` to suppress the compiler warning.

These changes should resolve the reported build failures and warnings.
